requirements.txt
altair
streamlit
matplotlib
Pillow
python-dotenv
pandas
google-genai
numpy
